+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/extensions/config/install/commit/InstallDumpApiPreCommitTask.kt
@@ -39,13 +39,13 @@ constructor(
             val installApiCheckPreCommitTask =
                 project.tasks.register<InstallDumpApiPreCommitTask>(name)
 
-            installApiCheckPreCommitTask.configure {
-                finalizedBy(WriteFilePreCommitTask.getTask(project))
+            installApiCheckPreCommitTask.configure { task ->
+                task.finalizedBy(WriteFilePreCommitTask.getTask(project))
             }
 
             project.tasks
                 .namedLazily<InstallPreCommitTask>(InstallPreCommitTask.name)
-                .configureEach { dependsOn(installApiCheckPreCommitTask) }
+                .configureEach { task -> task.dependsOn(installApiCheckPreCommitTask) }
         }
     }
 }