+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/extensions/kotlin/android/library/_internal/configureAndroidLibrary.kt
@@ -44,7 +44,7 @@ internal fun configureAndroidLibrary(project: Project) {
             defaultConfig.minSdk = androidState.minSdk
             namespace = project.calculateAndroidNamespace(androidState.namespace)
 
-            sourceSets.all { configDefaultAndroidSourceSets() }
+            sourceSets.all { set -> set.configDefaultAndroidSourceSets() }
             project.configureAndroidBuildFeatures(this)
         }
 
@@ -70,8 +70,8 @@ internal fun configureKotlinAndroidLibraryRawConfig(project: Project) {
 }
 
 private fun KotlinProjectExtension.configureAndroidDependencies() {
-    sourceSets.named("main") { dependencies { configureMainDependencies() } }
-    sourceSets.named("test") { dependencies { configureTestDependencies() } }
+    sourceSets.named("main") { set -> set.dependencies { configureMainDependencies() } }
+    sourceSets.named("test") { set -> set.dependencies { configureTestDependencies() } }
 }
 
 private fun KotlinDependencyHandler.configureMainDependencies() {